requirements.txt
deepgram-sdk
pynput
pystray
Pillow
pyaudio
PySide6
python-dotenv
